[WIP] C++17 Support (#535)
* C++17 Support

* C++17 compilation instructions

+++ include/BlockchainExplorerData.h
@@ -41,7 +41,7 @@ struct TransactionOutputDetails {
 
 struct TransactionOutputReferenceDetails {
   Crypto::Hash transactionHash;
-  size_t number;
+  uint64_t number;
 };
 
 struct BaseInputDetails {
