+++ src/Common/BlockingQueue.h
@@ -29,7 +29,7 @@ public:
 
   typedef BlockingQueue<T, Container> ThisType;
 
-  BlockingQueue(size_t maxSize = 1) : 
+  BlockingQueue(uint64_t maxSize = 1) : 
     m_maxSize(maxSize), m_closed(false) {}
 
   template <typename TT>
@@ -85,18 +85,18 @@ public:
     }
   }
 
-  size_t size() {
+  uint64_t size() {
     std::unique_lock<std::mutex> lk(m_mutex);
     return m_queue.size();
   }
 
-  size_t capacity() const {
+  uint64_t capacity() const {
     return m_maxSize;
   }
 
 private:
 
-  const size_t m_maxSize;
+  const uint64_t m_maxSize;
   Container m_queue;
   bool m_closed;
   
@@ -109,7 +109,7 @@ template <typename QueueT>
 class GroupClose {
 public:
 
-  GroupClose(QueueT& queue, size_t groupSize)
+  GroupClose(QueueT& queue, uint64_t groupSize)
     : m_queue(queue), m_count(groupSize) {}
 
   void close() {
@@ -121,7 +121,7 @@ public:
 
 private:
 
-  std::atomic<size_t> m_count;
+  std::atomic<uint64_t> m_count;
   QueueT& m_queue;
 
 };
