+++ src/Common/ConsoleHandler.cpp
@@ -199,7 +199,7 @@ namespace Common {
     
     std::stringstream ss;
 
-    size_t maxlen = std::max_element(m_handlers.begin(), m_handlers.end(), [](
+    uint64_t maxlen = std::max_element(m_handlers.begin(), m_handlers.end(), [](
       CommandHandlersMap::const_reference& a, CommandHandlersMap::const_reference& b) { 
         return a.first.size() < b.first.size(); })->first.size();
