+++ src/Common/IInputStream.h
@@ -17,6 +17,7 @@
 
 #pragma once
 
+#include <cstdint>
 #include <cstddef>
 
 namespace Common {
@@ -24,7 +25,7 @@ namespace Common {
 class IInputStream {
 public:
   virtual ~IInputStream() { }
-  virtual size_t readSome(void* data, size_t size) = 0;
+  virtual uint64_t readSome(void* data, uint64_t size) = 0;
 };
 
 }
