+++ src/Common/IOutputStream.h
@@ -17,6 +17,7 @@
 
 #pragma once
 
+#include <cstdint>
 #include <cstddef>
 
 namespace Common {
@@ -24,7 +25,7 @@ namespace Common {
 class IOutputStream {
 public:
   virtual ~IOutputStream() { }
-  virtual size_t writeSome(const void* data, size_t size) = 0;
+  virtual uint64_t writeSome(const void* data, uint64_t size) = 0;
 };
 
 }
