+++ src/Common/StdInputStream.cpp
@@ -22,7 +22,7 @@ namespace Common {
 StdInputStream::StdInputStream(std::istream& in) : in(in) {
 }
 
-size_t StdInputStream::readSome(void* data, size_t size) {
+uint64_t StdInputStream::readSome(void* data, uint64_t size) {
   in.read(static_cast<char*>(data), size);
   return in.gcount();
 }
