+++ src/Common/StringTools.cpp
@@ -44,7 +44,7 @@ const uint8_t characterValues[256] = {
 
 }
 
-std::string asString(const void* data, size_t size) {
+std::string asString(const void* data, uint64_t size) {
   return std::string(static_cast<const char*>(data), size);
 }
 
@@ -75,7 +75,7 @@ bool fromHex(char character, uint8_t& value) {
   return true;
 }
 
-size_t fromHex(const std::string& text, void* data, size_t bufferSize) {
+uint64_t fromHex(const std::string& text, void* data, uint64_t bufferSize) {
   if ((text.size() & 1) != 0) {
     throw std::runtime_error("fromHex: invalid string size");
   }
@@ -84,14 +84,14 @@ size_t fromHex(const std::string& text, void* data, size_t bufferSize) {
     throw std::runtime_error("fromHex: invalid buffer size");
   }
 
-  for (size_t i = 0; i < text.size() >> 1; ++i) {
+  for (uint64_t i = 0; i < text.size() >> 1; ++i) {
     static_cast<uint8_t*>(data)[i] = fromHex(text[i << 1]) << 4 | fromHex(text[(i << 1) + 1]);
   }
 
   return text.size() >> 1;
 }
 
-bool fromHex(const std::string& text, void* data, size_t bufferSize, size_t& size) {
+bool fromHex(const std::string& text, void* data, uint64_t bufferSize, uint64_t& size) {
   if ((text.size() & 1) != 0) {
     return false;
   }
@@ -100,7 +100,7 @@ bool fromHex(const std::string& text, void* data, size_t bufferSize, size_t& siz
     return false;
   }
 
-  for (size_t i = 0; i < text.size() >> 1; ++i) {
+  for (uint64_t i = 0; i < text.size() >> 1; ++i) {
     uint8_t value1;
     if (!fromHex(text[i << 1], value1)) {
       return false;
@@ -124,7 +124,7 @@ std::vector<uint8_t> fromHex(const std::string& text) {
   }
 
   std::vector<uint8_t> data(text.size() >> 1);
-  for (size_t i = 0; i < data.size(); ++i) {
+  for (uint64_t i = 0; i < data.size(); ++i) {
     data[i] = fromHex(text[i << 1]) << 4 | fromHex(text[(i << 1) + 1]);
   }
 
@@ -136,7 +136,7 @@ bool fromHex(const std::string& text, std::vector<uint8_t>& data) {
     return false;
   }
 
-  for (size_t i = 0; i < text.size() >> 1; ++i) {
+  for (uint64_t i = 0; i < text.size() >> 1; ++i) {
     uint8_t value1;
     if (!fromHex(text[i << 1], value1)) {
       return false;
@@ -153,9 +153,9 @@ bool fromHex(const std::string& text, std::vector<uint8_t>& data) {
   return true;
 }
 
-std::string toHex(const void* data, size_t size) {
+std::string toHex(const void* data, uint64_t size) {
   std::string text;
-  for (size_t i = 0; i < size; ++i) {
+  for (uint64_t i = 0; i < size; ++i) {
     text += "0123456789abcdef"[static_cast<const uint8_t*>(data)[i] >> 4];
     text += "0123456789abcdef"[static_cast<const uint8_t*>(data)[i] & 15];
   }
@@ -163,8 +163,8 @@ std::string toHex(const void* data, size_t size) {
   return text;
 }
 
-void toHex(const void* data, size_t size, std::string& text) {
-  for (size_t i = 0; i < size; ++i) {
+void toHex(const void* data, uint64_t size, std::string& text) {
+  for (uint64_t i = 0; i < size; ++i) {
     text += "0123456789abcdef"[static_cast<const uint8_t*>(data)[i] >> 4];
     text += "0123456789abcdef"[static_cast<const uint8_t*>(data)[i] & 15];
   }
@@ -172,7 +172,7 @@ void toHex(const void* data, size_t size, std::string& text) {
 
 std::string toHex(const std::vector<uint8_t>& data) {
   std::string text;
-  for (size_t i = 0; i < data.size(); ++i) {
+  for (uint64_t i = 0; i < data.size(); ++i) {
     text += "0123456789abcdef"[data[i] >> 4];
     text += "0123456789abcdef"[data[i] & 15];
   }
@@ -181,14 +181,14 @@ std::string toHex(const std::vector<uint8_t>& data) {
 }
 
 void toHex(const std::vector<uint8_t>& data, std::string& text) {
-  for (size_t i = 0; i < data.size(); ++i) {
+  for (uint64_t i = 0; i < data.size(); ++i) {
     text += "0123456789abcdef"[data[i] >> 4];
     text += "0123456789abcdef"[data[i] & 15];
   }
 }
 
 std::string extract(std::string& text, char delimiter) {
-  size_t delimiterPosition = text.find(delimiter);
+  uint64_t delimiterPosition = text.find(delimiter);
   std::string subText;
   if (delimiterPosition != std::string::npos) {
     subText = text.substr(0, delimiterPosition);
@@ -200,8 +200,8 @@ std::string extract(std::string& text, char delimiter) {
   return subText;
 }
 
-std::string extract(const std::string& text, char delimiter, size_t& offset) {
-  size_t delimiterPosition = text.find(delimiter, offset);
+std::string extract(const std::string& text, char delimiter, uint64_t& offset) {
+  uint64_t delimiterPosition = text.find(delimiter, offset);
   if (delimiterPosition != std::string::npos) {
     offset = delimiterPosition + 1;
     return text.substr(offset, delimiterPosition);
@@ -225,10 +225,10 @@ bool is_base64(unsigned char c) {
 }
 
 std::string base64Decode(std::string const& encoded_string) {
-  size_t in_len = encoded_string.size();
-  size_t i = 0;
-  size_t j = 0;
-  size_t in_ = 0;
+  uint64_t in_len = encoded_string.size();
+  uint64_t i = 0;
+  uint64_t j = 0;
+  uint64_t in_ = 0;
   unsigned char char_array_4[4], char_array_3[3];
   std::string ret;
 
@@ -272,7 +272,7 @@ bool loadFileToString(const std::string& filepath, std::string& buf) {
     fstream.exceptions(std::ifstream::failbit | std::ifstream::badbit);
     fstream.open(filepath, std::ios_base::binary | std::ios_base::in | std::ios::ate);
 
-    size_t fileSize = static_cast<size_t>(fstream.tellg());
+    uint64_t fileSize = static_cast<uint64_t>(fstream.tellg());
     buf.resize(fileSize);
 
     if (fileSize > 0)  {
