+++ src/Serialization/BinaryInputStreamSerializer.cpp
@@ -48,7 +48,7 @@ bool BinaryInputStreamSerializer::beginObject(Common::StringView name) {
 void BinaryInputStreamSerializer::endObject() {
 }
 
-bool BinaryInputStreamSerializer::beginArray(size_t& size, Common::StringView name) {
+bool BinaryInputStreamSerializer::beginArray(uint64_t& size, Common::StringView name) {
   readVarintAs<uint64_t>(stream, size);
   return true;
 }
@@ -130,7 +130,7 @@ bool BinaryInputStreamSerializer::operator()(std::string& value, Common::StringV
   return true;
 }
 
-bool BinaryInputStreamSerializer::binary(void* value, size_t size, Common::StringView name) {
+bool BinaryInputStreamSerializer::binary(void* value, uint64_t size, Common::StringView name) {
   checkedRead(static_cast<char*>(value), size);
   return true;
 }
@@ -145,7 +145,7 @@ bool BinaryInputStreamSerializer::operator()(double& value, Common::StringView n
   return false;
 }
 
-void BinaryInputStreamSerializer::checkedRead(char* buf, size_t size) {
+void BinaryInputStreamSerializer::checkedRead(char* buf, uint64_t size) {
   read(stream, buf, size);
 }
