+++ src/Serialization/BlockchainExplorerDataSerialization.cpp
@@ -138,24 +138,24 @@ void serialize(TransactionDetails& transaction, ISerializer& serializer) {
 
   //serializer(transaction.signatures, "signatures");
   if (serializer.type() == ISerializer::OUTPUT) {
-    std::vector<std::pair<size_t, Crypto::Signature>> signaturesForSerialization;
+    std::vector<std::pair<uint64_t, Crypto::Signature>> signaturesForSerialization;
     signaturesForSerialization.reserve(transaction.signatures.size());
-    size_t ctr = 0;
+    uint64_t ctr = 0;
     for (const auto& signaturesV : transaction.signatures) {
       for (auto signature : signaturesV) {
         signaturesForSerialization.emplace_back(ctr, std::move(signature));
       }
       ++ctr;
     }
-    size_t size = transaction.signatures.size();
+    uint64_t size = transaction.signatures.size();
     serializer(size, "signaturesSize");
     serializer(signaturesForSerialization, "signatures");
   } else {
-    size_t size = 0;
+    uint64_t size = 0;
     serializer(size, "signaturesSize");
     transaction.signatures.resize(size);
 
-    std::vector<std::pair<size_t, Crypto::Signature>> signaturesForSerialization;
+    std::vector<std::pair<uint64_t, Crypto::Signature>> signaturesForSerialization;
     serializer(signaturesForSerialization, "signatures");
 
     for (const auto& signatureWithIndex : signaturesForSerialization) {
