+++ src/Serialization/SerializationTools.h
@@ -87,14 +87,14 @@ void loadFromJsonValue(T& v, const Common::JsonValue& js) {
 
 template <typename T>
 void loadFromJsonValue(std::vector<T>& v, const Common::JsonValue& js) {
-  for (size_t i = 0; i < js.size(); ++i) {
+  for (uint64_t i = 0; i < js.size(); ++i) {
     v.push_back(Common::getValueAs<T>(js[i]));
   }
 }
 
 template <>
 inline void loadFromJsonValue(AddressBook &v, const Common::JsonValue &js) {
-  for (size_t i = 0; i < js.size(); ++i) {
+  for (uint64_t i = 0; i < js.size(); ++i) {
     AddressBookEntry type;
     loadFromJsonValue(type, js[i]);
     v.push_back(type);
@@ -103,7 +103,7 @@ inline void loadFromJsonValue(AddressBook &v, const Common::JsonValue &js) {
 
 template <typename T>
 void loadFromJsonValue(std::list<T>& v, const Common::JsonValue& js) {
-  for (size_t i = 0; i < js.size(); ++i) {
+  for (uint64_t i = 0; i < js.size(); ++i) {
     v.push_back(Common::getValueAs<T>(js[i]));
   }
 }
