+++ src/zedwallet/Menu.h
@@ -24,4 +24,4 @@ std::string getAction(Config &config);
 void mainLoop(std::shared_ptr<WalletInfo> walletInfo, CryptoNote::INode &node);
 
 template<typename T>
-void printCommands(const std::vector<T> &commands, int offset = 0);
+void printCommands(const std::vector<T> &commands, size_t offset = 0);
